Add all symlinks to Extra-source-files
Missing a symlink for Data/ByteString/Short/Internal.hs as well

The Extra-source-files is needed so that cabal sdist is guaranteed to pick
up the conditionally compiled bytestring builder.   Otherwise, these
files will get omitted if cabal configure found a version of ghc-7.8.

It'd be nice if a future version of cabal sdist would look at both sides
of a conditional statement regardless of what the condition actually
evaluates to.