Make Setup.hs compatible with Cabal 1.8 (GHC 6.12)
Making Setup.hs compatible with Cabal 1.6 is not easy, because
isSpecificVersion is not available. That would seem to imply that
that Setup.hs would need to refer to the ThisVersion constructor
directly, which is now deprecated.

And, I haven't found a way to use CPP to make Setup.hs that compiles
correctly across a wider versions of Cabal.