Adjust dependencies
Hide bytestring-builder behind an if impl(ghc < 7.8) to avoid downloading and
compiling bytestring-builder when it probably isn't needed.  This assumes
that GHC 7.8 users have at least bytestring-0.10.4,   and haven't installed
an older version of bytestring for some reason,  which should be a pretty
safe assumption.

Also,  text 1.1 is out now.